+++ src/client.rs
@@ -529,7 +529,7 @@ mod tests {
                     name: "test-wait-created-busybox".to_owned(),
                     image: Some("busybox:latest".to_owned()),
                     image_pull_policy: Some("IfNotPresent".to_owned()),
-                    command: vec!["sleep".into(), "infinity".into()],
+                    command: Some(vec!["sleep".into(), "infinity".into()]),
                     ..Container::default()
                 }],
                 termination_grace_period_seconds: Some(1),
@@ -618,7 +618,7 @@ mod tests {
         let mut match_labels: BTreeMap<String, String> = BTreeMap::new();
         match_labels.insert("app".to_owned(), "busybox".to_owned());
         let label_selector: LabelSelector = LabelSelector {
-            match_labels: match_labels.clone(),
+            match_labels: Some(match_labels.clone()),
             ..LabelSelector::default()
         };
         let no_pods: Vec<Pod> = client
@@ -630,7 +630,7 @@ mod tests {
         let pod_to_wait_for: Pod = Pod {
             metadata: ObjectMeta {
                 name: Some("pod-to-be-listed".to_owned()),
-                labels: match_labels.clone(),
+                labels: Some(match_labels.clone()),
                 ..ObjectMeta::default()
             },
             spec: Some(PodSpec {
@@ -638,7 +638,7 @@ mod tests {
                     name: "test-wait-created-busybox".to_owned(),
                     image: Some("busybox:latest".to_owned()),
                     image_pull_policy: Some("IfNotPresent".to_owned()),
-                    command: vec!["sleep".into(), "infinity".into()],
+                    command: Some(vec!["sleep".into(), "infinity".into()]),
                     ..Container::default()
                 }],
                 termination_grace_period_seconds: Some(1),