+++ src/lib.rs
@@ -23,8 +23,10 @@ pub mod pod_utils;
 pub mod product_config_utils;
 pub mod reconcile;
 pub mod role_utils;
+pub mod scheduler;
 pub mod status;
 pub mod utils;
 pub mod validation;
+pub mod versioning;
 
 pub use crate::crd::CustomResourceExt;