+++ src/role_utils.rs
@@ -81,7 +81,7 @@
 //! Each resource can have more operator specific labels.
 
 use crate::error::OperatorResult;
-use crate::{krustlet, label_selector, labels};
+use crate::{krustlet, labels};
 
 use std::collections::{BTreeMap, HashMap};
 
@@ -163,7 +163,6 @@ pub struct RoleGroup<T> {
     #[serde(flatten)]
     pub config: Option<CommonConfiguration<T>>,
     pub replicas: Option<u16>,
-    #[schemars(schema_with = "label_selector::schema")]
     pub selector: Option<LabelSelector>,
 }
 
@@ -174,7 +173,7 @@ pub async fn find_nodes_that_fit_selectors<T>(
     client: &Client,
     namespace: Option<String>,
     role: &Role<T>,
-) -> OperatorResult<HashMap<String, (Vec<Node>, Option<u16>)>>
+) -> OperatorResult<HashMap<String, EligibleNodesAndReplicas>>
 where
     T: Serialize,
 {
@@ -190,14 +189,27 @@ where
             group_name,
             nodes
         );
-        found_nodes.insert(group_name.clone(), (nodes, role_group.replicas));
+        found_nodes.insert(
+            group_name.clone(),
+            EligibleNodesAndReplicas {
+                nodes,
+                replicas: role_group.replicas,
+            },
+        );
     }
     Ok(found_nodes)
 }
 
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct EligibleNodesAndReplicas {
+    pub nodes: Vec<Node>,
+    pub replicas: Option<u16>,
+}
+
 /// Type to avoid clippy warnings
-/// HashMap<`NameOfRole`, HashMap<`NameOfRoleGroup`, (Vec<`Node`>, Option<`Replicas`>)>>
-pub type EligibleNodesForRoleAndGroup = HashMap<String, HashMap<String, (Vec<Node>, Option<u16>)>>;
+/// HashMap<`NameOfRole`, HashMap<`NameOfRoleGroup`, EligibleNodesAndReplicas(Vec<`Node`>, Option<`Replicas`>)>>
+pub type EligibleNodesForRoleAndGroup = HashMap<String, HashMap<String, EligibleNodesAndReplicas>>;
 
 /// Return a list of eligible nodes and the provided replica count for each role and group
 /// combination. Required to delete excess pods that do not match any node, selector description
@@ -211,17 +223,17 @@ pub fn list_eligible_nodes_for_role_and_group(
 ) -> Vec<(Vec<Node>, LabelOptionalValueMap, Option<u16>)> {
     let mut eligible_nodes_for_role_and_group = vec![];
     for (role, eligible_nodes_for_role) in eligible_nodes {
-        for (group_name, (eligible_nodes, replicas)) in eligible_nodes_for_role {
+        for (group_name, eligible_nodes) in eligible_nodes_for_role {
             trace!(
                 "Adding {} nodes to eligible node list for role [{}] and group [{}].",
-                eligible_nodes.len(),
+                eligible_nodes.nodes.len(),
                 role,
                 group_name
             );
             eligible_nodes_for_role_and_group.push((
-                eligible_nodes.clone(),
+                eligible_nodes.nodes.clone(),
                 get_role_and_group_labels(role, group_name),
-                *replicas,
+                eligible_nodes.replicas,
             ))
         }
     }
@@ -256,7 +268,7 @@ mod tests {
             - node_1
             - node_2
           group2:
-            - node_3"
+            - node_3
     "#
     )]
     #[case::two_roles(
@@ -319,7 +331,13 @@ mod tests {
                     collected_nodes.push(node);
                 }
                 // replicas (0) does not affect here
-                group_map.insert(group_name.clone(), (collected_nodes, Some(0u16)));
+                group_map.insert(
+                    group_name.clone(),
+                    EligibleNodesAndReplicas {
+                        nodes: collected_nodes,
+                        replicas: Some(0u16),
+                    },
+                );
             }
             eligible_nodes.insert(role_name.clone(), group_map);
         }